+++ api/api/Exchanges/Bybit/BybitService.cs
@@ -11,21 +11,15 @@ public class BybitService : IBybitService
     private const string SubMembersEndpoint = "/v5/user/submembers";
     private const int RecvWindow = 5000;
 
-    private readonly string _tradingBaseUrl;
     private readonly string _accountBaseUrl;
     private readonly ILogger<BybitService> _logger;
 
     public BybitService(IOptions<BybitSettings> settings, ILogger<BybitService> logger)
     {
-        _tradingBaseUrl = settings.Value.TradingBaseUrl;
-        _accountBaseUrl = settings.Value.AccountBaseUrl;
+        _accountBaseUrl = "https://api.bybit.com";
         _logger = logger;
-        _logger.LogInformation("BybitService initialised — trading: {TradingUrl} | accounts: {AccountUrl}",
-            _tradingBaseUrl, _accountBaseUrl);
     }
 
-    // Bybit webhook signature: HMAC-SHA256(webhookSecret, timestamp + rawBody), hex-encoded.
-    // Header names: X-Bybit-Timestamp, X-Bybit-Signature.
     public bool ValidateWebhookSignature(string rawBody, string signature, string timestamp, string webhookSecret)
     {
         try
@@ -52,7 +46,13 @@ public class BybitService : IBybitService
         try
         {
             var timestamp = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds().ToString();
-            var signature = BuildApiSignature(apiKey, apiSecret, timestamp, string.Empty);
+            var paramStr = $"{timestamp}{apiKey}{RecvWindow}";
+            var keyBytes = Encoding.UTF8.GetBytes(apiSecret);
+            var paramBytes = Encoding.UTF8.GetBytes(paramStr);
+
+            using var hmac = new HMACSHA256(keyBytes);
+            var hashBytes = hmac.ComputeHash(paramBytes);
+            var signature = Convert.ToHexString(hashBytes).ToLowerInvariant();
 
             var response = await _accountBaseUrl
                 .AppendPathSegment(SubMembersEndpoint)
@@ -75,17 +75,5 @@ public class BybitService : IBybitService
             _logger.LogError(ex, "Error fetching Bybit sub-accounts");
             throw;
         }
-    }
-
-    // Bybit REST API signature: HMAC-SHA256(apiSecret, timestamp + apiKey + recvWindow + queryString).
-    private static string BuildApiSignature(string apiKey, string apiSecret, string timestamp, string queryString)
-    {
-        var paramStr = $"{timestamp}{apiKey}{RecvWindow}{queryString}";
-        var keyBytes = Encoding.UTF8.GetBytes(apiSecret);
-        var paramBytes = Encoding.UTF8.GetBytes(paramStr);
-
-        using var hmac = new HMACSHA256(keyBytes);
-        var hashBytes = hmac.ComputeHash(paramBytes);
-        return Convert.ToHexString(hashBytes).ToLowerInvariant();
     }
 }
