+++ api/unit-tests/ExchangesTests/Bybit/BybitSettingsTests.cs
@@ -5,28 +5,18 @@ namespace unit_tests.ExchangesTests.Bybit;
 public class BybitSettingsTests
 {
     [Fact]
-    public void TradingBaseUrl_WhenUseTestnetIsFalse_ShouldReturnProduction()
+    public void UseTestnet_DefaultsToFalse()
     {
-        var settings = new BybitSettings { UseTestnet = false };
+        var settings = new BybitSettings();
 
-        settings.TradingBaseUrl.Should().Be("https://api.bybit.com");
+        settings.UseTestnet.Should().BeFalse();
     }
 
     [Fact]
-    public void TradingBaseUrl_WhenUseTestnetIsTrue_ShouldReturnTestnet()
+    public void UseTestnet_CanBeSetToTrue()
     {
         var settings = new BybitSettings { UseTestnet = true };
 
-        settings.TradingBaseUrl.Should().Be("https://api-testnet.bybit.com");
-    }
-
-    [Fact]
-    public void AccountBaseUrl_ShouldAlwaysBeProduction()
-    {
-        var production = new BybitSettings { UseTestnet = false };
-        var testnet = new BybitSettings { UseTestnet = true };
-
-        production.AccountBaseUrl.Should().Be("https://api.bybit.com");
-        testnet.AccountBaseUrl.Should().Be("https://api.bybit.com");
+        settings.UseTestnet.Should().BeTrue();
     }
 }
